+++ spec/dummy_app/app/active_record/two_level/namespaced.rb
@@ -1,0 +1,7 @@
+module TwoLevel
+  module Namespaced
+    def self.table_name_prefix
+      'two_level_namespaced_'
+    end
+  end
+end